# openapi.v1.json
{
  "components": {
    "schemas": {
...
      "BatchRequest": {
        "additionalProperties": false,
        "properties": {
          "callback_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional https URL that receives a signed JobResponse per finished job",
            "title": "Callback Url"
          },
          "cas": {
            "default": [],
            "description": "CAs to prefetch. Use `items` instead to pass hints; both may be combined (at most 50 in total, at least one)",
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Cas",
            "type": "array"
          },
          "depth": {
            "default": "basic",
            "enum": [
              "basic",
              "full"
            ],
            "title": "Depth",
            "type": "string"
          },
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BatchRequestItem"
            },
            "maxItems": 50,
            "title": "Items",
            "type": "array"
          }
        },
        "title": "BatchRequest",
        "type": "object"
      },
      "BatchRequestItem": {
        "additionalProperties": false,
        "properties": {
          "ca": {
            "title": "Ca",
            "type": "string"
          },
          "hints": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenHints"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ca"
        ],
        "title": "BatchRequestItem",
        "type": "object"
      },
      "BatchResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchItem"
            },
            "title": "Items",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "items",
          "request_id"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
...
      "FlagEntry": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "severity": {
            "enum": [
              "info",
              "warn",
              "high"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "severity",
          "description"
        ],
        "title": "FlagEntry",
        "type": "object"
      },
      "Freshness": {
        "additionalProperties": false,
        "properties": {
          "age_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age S"
          },
          "analyzed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed At"
          },
          "from_cache": {
            "default": false,
            "title": "From Cache",
            "type": "boolean"
          },
          "max_age_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Age S"
          }
        },
        "title": "Freshness",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobResponse": {
        "additionalProperties": false,
        "properties": {
          "ca": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ca"
          },
          "depth": {
            "anyOf": [
              {
                "enum": [
                  "basic",
                  "full"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depth"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "request_id"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "MetaResponse": {
        "additionalProperties": false,
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/TaxonomyEntry"
            },
            "title": "Categories",
            "type": "array"
          },
          "depths": {
            "items": {
              "enum": [
                "basic",
                "full"
              ],
              "type": "string"
            },
            "title": "Depths",
            "type": "array"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "flags": {
            "items": {
              "$ref": "#/components/schemas/FlagEntry"
            },
            "title": "Flags",
            "type": "array"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "service_version": {
            "title": "Service Version",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/Versions"
          }
        },
        "required": [
          "schema_version",
          "service_version",
          "versions",
          "categories",
          "flags",
          "depths",
          "disclaimer"
        ],
        "title": "MetaResponse",
        "type": "object"
      },
...
      "TaxonomyEntry": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "description"
        ],
        "title": "TaxonomyEntry",
        "type": "object"
      },
...
      "TokenRequest": {
        "additionalProperties": false,
        "description": "Optional body of POST /v1/tokens/{ca} (the query parameters are those of GET).",
        "properties": {
          "hints": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenHints"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "TokenRequest",
        "type": "object"
      },
      "TokenResponse": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Analysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "ca": {
            "title": "Ca",
            "type": "string"
          },
          "depth": {
            "enum": [
              "basic",
              "full"
            ],
            "title": "Depth",
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UpstreamError"
            },
            "title": "Errors",
            "type": "array"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness",
            "default": {
              "from_cache": false
            }
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "stale_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Analysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "complete",
              "partial",
              "pending",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "ca",
          "status",
          "depth",
          "request_id"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UpstreamError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "code"
        ],
        "title": "UpstreamError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Versions": {
        "additionalProperties": false,
        "properties": {
          "known_coins": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known Coins"
          },
          "lexicon": {
            "title": "Lexicon",
            "type": "string"
          },
          "models": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Models",
            "type": "object"
          },
          "rules": {
            "title": "Rules",
            "type": "string"
          }
        },
        "required": [
          "rules",
          "lexicon"
        ],
        "title": "Versions",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Explains what a pump.fun token means, given its contract address. Deterministic analysis of name, ticker, image and linked X/Twitter content.",
    "title": "TokenSage",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Poll an analysis job",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/meta": {
      "get": {
        "operationId": "meta_v1_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Versions, taxonomy and flag codes",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/tokens/{ca}": {
      "get": {
        "operationId": "get_token_v1_tokens__ca__get",
        "parameters": [
          {
            "in": "path",
            "name": "ca",
            "required": true,
            "schema": {
              "title": "Ca",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "basic",
                    "full"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Depth"
            }
          },
          {
            "in": "query",
            "name": "wait",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 25,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wait"
            }
          },
          {
            "in": "query",
            "name": "max_age",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 604800,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Age"
            }
          },
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          },
          {
            "description": "comma list of optional parts to keep: evidence, raw. Omit to get everything; a part left out of the list is dropped (evidence -> [], raw -> empty)",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "comma list of optional parts to keep: evidence, raw. Omit to get everything; a part left out of the list is dropped (evidence -> [], raw -> empty)",
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Analysis pending; poll again."
          },
          "400": {
            "description": "invalid_ca"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "token_not_found (never when hints with metadata were given)"
          },
          "422": {
            "description": "not_a_token_mint | not_pumpfun"
          },
          "429": {
            "description": "rate_limited | quota_exceeded"
          },
          "503": {
            "description": "overloaded"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analyze a pump.fun token by contract address",
        "tags": [
          "v1"
        ]
      },
      "post": {
        "description": "Same as GET /v1/tokens/{ca} (same query parameters and responses), with an optional JSON body carrying `hints`. With hints the metadata fetch is skipped, and a mint not yet visible on-chain is analysed from the hints instead of returning 404.",
        "operationId": "post_token_v1_tokens__ca__post",
        "parameters": [
          {
            "in": "path",
            "name": "ca",
            "required": true,
            "schema": {
              "title": "Ca",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "basic",
                    "full"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Depth"
            }
          },
          {
            "in": "query",
            "name": "wait",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 25,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wait"
            }
          },
          {
            "in": "query",
            "name": "max_age",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 604800,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Age"
            }
          },
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          },
          {
            "description": "comma list of optional parts to keep: evidence, raw. Omit to get everything; a part left out of the list is dropped (evidence -> [], raw -> empty)",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "comma list of optional parts to keep: evidence, raw. Omit to get everything; a part left out of the list is dropped (evidence -> [], raw -> empty)",
              "title": "Include"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TokenRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Analysis pending; poll again."
          },
          "400": {
            "description": "invalid_ca"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "token_not_found (never when hints with metadata were given)"
          },
          "422": {
            "description": "not_a_token_mint | not_pumpfun"
          },
          "429": {
            "description": "rate_limited | quota_exceeded"
          },
          "503": {
            "description": "overloaded"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analyze a token, passing metadata hints you already have",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/tokens:batch": {
      "post": {
        "operationId": "batch_v1_tokens_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Prefetch up to 50 CAs",
        "tags": [
          "v1"
        ]
      }
    }
  }
}